requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.2
nltk==3.9
gensim==4.3.2
pyLDAvis==3.4.1
matplotlib==3.7.2
wordcloud==1.9.3
transformers==4.34.0
fpdf2==2.7.8
python-docx==1.1.0
pdfplumber==0.10.3
langdetect==1.0.9
googletrans==4.0.0rc1
sentence-transformers==2.2.2
scipy==1.10.1
seaborn==0.12.2
plotly==5.18.0
beautifulsoup4==4.12.3
lxml==4.9.3
ebooklib==0.18
Pillow==9.5.0
pytesseract==0.3.10
python-pptx==0.6.21
striprtf==0.0.28
openpyxl==3.1.2
spacy==3.7.2
networkx==3.1
pyvis==0.3.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl